WL-225 Added functionality which enables forum notifications via e-mail.
This is taken from the work started by Lydia Li at Stanford and improved by
David Horwitz at UCT on SAK-11740. Specifically, this commit is the same as the
revised-SAK-11740-26x.patch attached to that JIRA (by me) and is the work
patched against a 2.6.x version of the msgcntr tool. (msgcntr-2.7.x encompasses
SAK-11740 along with others, but does not appear to have progressed that far yet
- probably one for the next release of WebLearn! - so in the mean time, it seems
expedient to commit this work which single-mindedly addresses this issue alone).