WL-1054 Added how to alter forum notification settings as footer to e-mail.
As the e-mail is already HTML, the footer includes a hyperlink to take the user
to the corresponding forum. Unfortunately, due to the tool remembering state, if
the user is already logged in, they may not see the 'Watch' link. However, there
is no way to establish the reset URL by querying the API, so unfortunately
crafting this by hand might break in a future implementation. Thus the message
includes a hint that a tool reset may be required. (I also added a blank line of
padding above the body of the message post, as I think this improves the
presentation).