WL-1066 Better fix to suppress DEBUG output of forum e-mail notifications.
This fix (after reverting previous wonky "fix") simply removes the setting of
setDebug(true) that was occuring on the corresponding javax.mail.Session object.
There is no attempt to make this configurable via any sakai property (unlike
before!). To be honest there doesn't seem to be an obvious one in existence to
use for this purpose (?)