BSGAPP-12 WP5.1 extra web services
WL-1949 Stack trace when creating a new Forum.
WL-2020 Exception editing topic settings
WL-2021 When saving a message in messages don't load topic.
WL-2141 Make sure hibernate session stays with transaction.
WL-2225 Photo URL via forums API
WL-2249 Fix permission checks in direct/forums

When a message is created in the messages tool there isn't a topic associated with it. Now we check and don't fall over when this is the case.

Move DecoratedMessage from inner class to class

Fix to deleteMessage

Bugfix to forums.newtopic event

The problem with msgcntr seems to be that the hibernate session doesn't stick around with the transaction, which means that all the lazy loading doesn't work outside the current method. Adding the HibernateInterceptor fixes this.

Conflicts:
	messageforums-api/src/java/org/sakaiproject/api/app/messageforums/ui/UIPermissionsManager.java
	messageforums-component-impl/src/java/org/sakaiproject/component/app/messageforums/entity/ForumTopicEntityProviderImpl.java
	messageforums-component-impl/src/java/org/sakaiproject/component/app/messageforums/ui/DiscussionForumManagerImpl.java
	messageforums-component-impl/src/java/org/sakaiproject/component/app/messageforums/ui/UIPermissionsManagerImpl.java