WL-2099 Profile details appear with click.
Previously profile details appeared when hovering, this meant that when scrolling through a big thread you might get loads of profile details appearing then disappearing quickly. This changes it so you have to click on the person to get the details.

It also moves the JavaScript into a common file rather than having it duplicated across 3 files.