WL-2398 Move the filtering into the tool.
This is because the tool re-saves the message at the head of the thread when new posts are appended to the topic and this was causing it to becomes  escaped.
The duplication of code in the tool and impl needs to be fixed but not under this bug.

Also added support for adding a plain text header when revising a post through the /direct interface.