WL-3100 Remove roles from forums permissions
The forums tool does not correctly handle permissions for roles in the
forums settings page. As this is the case and the forums tool uses a
proprietary permissions management system, we have removed the roles
from being displayed in the dropdown list.

As far as we can tell the groups and special roles all work as expected.