WL-3561 Saving of a message when values aren’t passed
We now are much more robust in how we work out values when they aren’t passed to us. This patch is far from ideal as it means there are multiple ways to work the information out, but it allows the code to work for both the tool (without change) and from /direct.